+++ backend/src/app.module.ts
@@ -8,6 +8,7 @@ import { SettingsModule } from './settings/settings.module';
 import { HolidaysModule } from './holidays/holidays.module';
 import { WebPushModule } from './web-push/web-push.module';
 import { WorkCalendarModule } from './work-calendar/work-calendar.module';
+import { ProfileModule } from './profile/profile.module';
 const config = getConfig();
 
 const url = `${config.DATABASE_URL}`;
@@ -15,6 +16,7 @@ const url = `${config.DATABASE_URL}`;
 @Module({
   imports: [
     WorkCalendarModule,
+    ProfileModule,
     MailModule,
     HolidaysModule,
     DictionaryModule,
