+++ backend/src/profile/services/avatars/confluence-avatars.service.ts
