+++ backend/src/profile/services/avatars/default-avatars.service.ts
