+++ backend/src/profile/services/users.service.ts
@@ -1,6 +1,5 @@
 import * as crypto from 'crypto';
 import { Model } from 'mongoose';
-import { LoginModel } from '../models/login.model';
 import { UserModel } from '../models/user.model';
 import { UserEntity } from '../../entity/entities/user.entity.model';
 import { Injectable } from '@nestjs/common';
@@ -38,30 +37,6 @@ export class UsersService {
 
   async addUser(userInfo: UserModel): Promise<UserEntity> {
     const newUser = await this.userModel.create(userInfo);
-    return newUser.save();
-  }
-
-  async registration(userInfo: LoginModel): Promise<UserEntity> {
-    const data: UserModel = {
-      id: null,
-      username: userInfo.name,
-      location: null,
-      position: null,
-      projects: [],
-      whenCreated: null,
-      email: `${userInfo.username}@it2g.ru`,
-      telNumber: null,
-      physicalDeliveryOfficeName: null,
-      mailNickname: userInfo.username,
-      isAdmin: false,
-      hasMailing: false,
-      subdivision: null,
-      jobPosition: null,
-      authType: 'hash',
-      hashPassword: crypto.createHmac('sha256', userInfo.password).digest('hex')
-    };
-
-    const newUser = await this.userModel.create(data);
     return newUser.save();
   }
