+++ backend/src/work-calendar/services/ldap.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, OnApplicationShutdown, Logger } from '@nestjs/common';
 import { Config } from '../../config/config';
 import { LoginModel } from '../models/login.model';
-import { UserModel } from '../models/user.model';
 import * as ldap from 'ldapjs';
+import { UserModel } from '../../profile/models/user.model';
 @Injectable()
 export class LdapService implements OnApplicationShutdown {
   private readonly logger = new Logger('LdapService');
@@ -46,7 +46,7 @@ export class LdapService implements OnApplicationShutdown {
 
   private mapToSendOnClient(attributes: Array<{ type: string; data: string }>): UserModel {
     return {
-      id: null,
+      _id: null,
       username: this.getAttribute(attributes, 'cn'),
       location: this.getAttribute(attributes, 'l'),
       position: this.getAttribute(attributes, 'title'),
@@ -55,7 +55,6 @@ export class LdapService implements OnApplicationShutdown {
       telNumber: this.getAttribute(attributes, 'mobile'),
       physicalDeliveryOfficeName: this.getAttribute(attributes, 'physicalDeliveryOfficeName'),
       mailNickname: this.getAttribute(attributes, 'mailNickname'),
-      projects: [],
       isAdmin: false,
       hasMailing: true,
       subdivision: null,
